rename the functions!
