+++ test/Jsonfile/Set1.json
@@ -1,0 +1,206 @@
+{
+  "layer0": [
+    {
+      "AggCenter": [
+        197.0,
+        143.0
+      ],
+      "AggrObjects": [
+        {
+          "GUID": "object-749f49a9-4405-46ae-9091-6dd3822a1b72",
+          "Vertices": [
+            [
+              217.0,
+              263.0
+            ],
+            [
+              202.0,
+              270.0
+            ],
+            [
+              206.0,
+              241.0
+            ]
+          ],
+          "center": [
+            208.33333333333334,
+            258.0
+          ],
+          "nrVertices": 3,
+          "orientation": 45,
+          "radii": 35
+        },
+        {
+          "GUID": "object-5f08a039-2302-4be4-8779-79604687249b",
+          "Vertices": [
+            [
+              306.0,
+              84.0
+            ],
+            [
+              282.0,
+              112.0
+            ],
+            [
+              268.0,
+              64.0
+            ],
+            [
+              292.0,
+              36.0
+            ]
+          ],
+          "center": [
+            287.0,
+            74.0
+          ],
+          "nrVertices": 4,
+          "orientation": 45,
+          "radii": 80
+        },
+        {
+          "GUID": "object-c54e787a-9ae2-44f7-b663-c4692fdda0f7",
+          "Vertices": [
+            [
+              190.0,
+              161.0
+            ],
+            [
+              179.0,
+              180.0
+            ],
+            [
+              166.0,
+              163.0
+            ],
+            [
+              170.0,
+              135.0
+            ],
+            [
+              185.0,
+              133.0
+            ]
+          ],
+          "center": [
+            178.0,
+            154.4
+          ],
+          "nrVertices": 5,
+          "orientation": 45,
+          "radii": 50
+        }
+      ],
+      "AggregateObjectName": "AggregateObject0",
+      "rotation": 36,
+      "scale": 2,
+      "translation": [
+        1.8,
+        2,
+        5
+      ]
+    }
+  ],
+  "layer1": [
+    {
+      "AggCenter": [
+        72.0,
+        95.0
+      ],
+      "AggrObjects": [
+        {
+          "GUID": "object-6fe232a5-cd9f-44c0-9c55-bd14f1414db2",
+          "Vertices": [
+            [
+              78.0,
+              49.0
+            ],
+            [
+              72.0,
+              42.0
+            ],
+            [
+              78.0,
+              35.0
+            ]
+          ],
+          "center": [
+            76.0,
+            42.0
+          ],
+          "nrVertices": 3,
+          "orientation": 45,
+          "radii": 15
+        },
+        {
+          "GUID": "object-1be7993e-e60f-43b5-b5d3-5a3330d2d9fb",
+          "Vertices": [
+            [
+              90.0,
+              95.0
+            ],
+            [
+              84.0,
+              92.0
+            ],
+            [
+              84.0,
+              78.0
+            ],
+            [
+              92.0,
+              81.0
+            ]
+          ],
+          "center": [
+            87.5,
+            86.5
+          ],
+          "nrVertices": 4,
+          "orientation": 45,
+          "radii": 20
+        },
+        {
+          "GUID": "object-8b50f1da-da09-4b67-b187-10f90b33ccdb",
+          "Vertices": [
+            [
+              120.0,
+              166.0
+            ],
+            [
+              114.0,
+              166.0
+            ],
+            [
+              112.0,
+              155.0
+            ],
+            [
+              117.0,
+              148.0
+            ],
+            [
+              122.0,
+              154.0
+            ]
+          ],
+          "center": [
+            117.0,
+            157.8
+          ],
+          "nrVertices": 5,
+          "orientation": 45,
+          "radii": 20
+        }
+      ],
+      "AggregateObjectName": "AggregateObject1",
+      "rotation": 36,
+      "scale": 2,
+      "translation": [
+        1.8,
+        2,
+        5
+      ]
+    }
+  ]
+}
